Translate specific port kinds for transport node features.
If all of the ultimate sources and destinations of transport nodes are
of the same port kind (data port or event data port), then the transport
node features will be of that kind. If there is a mixture of data port
and event data port, then the features are translated as abstract
features.

If a specific port kind is identified for a transport node, then all
connections to and from that transport node will be port connections, If
a specific port kind is not identified, then they will be feature
connections.