Select Platform only by default.
Many FACE models are very large and when they are translated into AADL,
there are so many data classifiers that OSATE slows down significantly.
By defaulting to Platform only, less classifiers are produced for these
large models and OSATE performs better. It is also rare that the
conceptual or logical elements are needed in AADL.